processing/requirements.txt
connexion[flask]==3.0.5
swagger-ui-bundle==1.1.0
APScheduler==3.10.4
SQLAlchemy==2.0.25
requests==2.31.0
PyYAML==6.0.1
uvicorn==0.27.0
flask_cors==4.0.0
